app notifications constants
